test: factor stringToUint() into its own file
