+++ test/lib/stringToUint.js
@@ -1,0 +1,11 @@
+// Conversion copied from:
+// https://stackoverflow.com/questions/17191945/conversion-between-utf-8-arraybuffer-and-string#answer-17192845
+module.exports = function stringToUint(string) {
+  var string = unescape(encodeURIComponent(string)),
+    charList = string.split(''),
+    uintArray = [];
+  for (var i = 0; i < charList.length; i++) {
+    uintArray.push(charList[i].charCodeAt(0));
+  }
+  return new Uint8Array(uintArray);
+};
